# docs/api/openapi3.yaml
openapi: 3.0.1
info:
  title: sync
  description: ""
  version: 0.0.1
servers:
- url: ""
...
paths:
...
  /posts:
    get:
      tags:
      - post
      summary: Get Posts
      description: Get Posts
      operationId: GetPosts
      parameters:
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPosts:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
    post:
      tags:
      - post
      summary: Create Post
      description: Create Post
      operationId: CreatePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            examples:
              CreatePost:
                value: "{\"title\":\"title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                  ,\"content\":{\"text\":\"This is a post content.\",\"json\":\"{\\\
                  \"text\\\": \\\"This is a post content.\\\"}\",\"mediaIds\":[1,2]},\"\
                  tags\":[\"java\",\"spring\"],\"referencedPostIds\":[10,11],\"coverMediaId\"\
                  :\"100\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                CreatePost:
                  value: |-
                    {
                      "slug" : "slug"
                    }
  /posts/drafts:
    get:
      tags:
      - post
      summary: Get Draft Posts
      description: Get Draft Posts
      operationId: GetDraftPosts
      parameters:
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      - name: type
        in: query
        description: 게시글 타입
        required: false
        schema:
          type: string
      - name: scope
        in: query
        description: 게시글 공개 범위
        required: false
        schema:
          type: string
      - name: projectHandle
        in: query
        description: 프로젝트로 검색 범위 제한 (선택)
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetDraftPosts:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"DRAFT\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
...
  /posts/{postId}:
    delete:
      tags:
      - post
      summary: Delete Post
      description: Delete Post
      operationId: DeletePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - post
      summary: Update Post
      description: Update Post
      operationId: UpdatePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
            examples:
              UpdatePost:
                value: "{\"title\":\"Updated post title\",\"type\":\"LONG\",\"status\"\
                  :\"PUBLISHED\",\"content\":{\"text\":\"This is a post content.\"\
                  ,\"json\":\"{\\\"text\\\": \\\"This is a post content.\\\"}\",\"\
                  mediaIds\":[1]},\"tags\":[\"java\"],\"referencedPostIds\":[10,11],\"\
                  coverMediaId\":\"100\",\"removeCover\":false}"
      responses:
        "204":
          description: "204"
  /posts/{slug}:
    get:
      tags:
      - post
      summary: Get Post By Slug
      description: Get Post By Slug
      operationId: GetPostBySlug
      parameters:
      - name: slug
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostResponse"
              examples:
                GetPostBySlug:
                  value: "{\"summary\":{\"id\":1,\"slug\":\"test-slug\",\"title\"\
                    :\"Test Post Title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                    ,\"scope\":\"WORKSPACE\",\"accessLevel\":\"FULL\",\"author\":{\"\
                    handle\":\"john-doe\",\"name\":\"John Doe\",\"profileImageUrl\"\
                    :\"https://example.com/john.png\"},\"project\":{\"handle\":\"\
                    my-project\",\"name\":\"나의 프로젝트\",\"description\":\"프로젝트 설명\"\
                    ,\"website\":\"https://example.com\",\"isPublic\":true,\"joinPolicy\"\
                    :\"INVITE\",\"followerCount\":42,\"iconUrl\":\"https://example.com/icon.png\"\
                    },\"resolved\":false,\"isSeriesPost\":false,\"pinnedAt\":null,\"\
                    isAuthor\":false,\"canDelete\":false,\"canComment\":true,\"createdAt\"\
                    :\"1970-01-01T00:00:00Z\",\"updatedAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"likeCount\":1,\"liked\":true,\"commentCount\":1,\"bookmarked\"\
                    :true,\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false}],\"preview\":\"This is a preview\
                    \ of the post content.\",\"previewMedia\":[{\"id\":1,\"url\":\"\
                    https://example.com/image.png\",\"fileName\":\"image.png\",\"\
                    fileSize\":102400,\"mediaType\":\"image/png\"}],\"mediaCount\"\
                    :1,\"wordCount\":120,\"coverImageUrl\":\"https://example.com/cover.png\"\
                    },\"content\":{\"json\":\"Post Content\",\"media\":[{\"id\":1,\"\
                    url\":\"https://example.com/media.png\",\"fileName\":\"media.png\"\
                    ,\"fileSize\":102400,\"mediaType\":\"image/png\"}]}}"
...
  /posts/{slug}/referenced-by:
    get:
      tags:
      - post
      summary: Get Post Backlinks
      description: "이 게시글을 가리키는(역참조, backlink) 게시글 목록을 조회합니다."
      operationId: GetPostBacklinks
      parameters:
      - name: slug
        in: path
        description: 게시글 slug
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPostBacklinks:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
  /posts/{slug}/references:
    get:
      tags:
      - post
      summary: Get Post References
      description: 이 게시글이 가리키는 참조(forward reference) 목록을 조회합니다.
      operationId: GetPostReferences
      parameters:
      - name: slug
        in: path
        description: 게시글 slug
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostsResponse"
              examples:
                GetPostReferences:
                  value: "{\"posts\":[{\"id\":1,\"slug\":\"test-slug\",\"title\":\"\
                    Test Post Title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                    ,\"scope\":\"WORKSPACE\",\"accessLevel\":\"FULL\",\"author\":{\"\
                    handle\":\"john-doe\",\"name\":\"John Doe\",\"profileImageUrl\"\
                    :\"https://example.com/john.png\"},\"project\":{\"handle\":\"\
                    my-project\",\"name\":\"나의 프로젝트\",\"description\":\"프로젝트 설명\"\
                    ,\"website\":\"https://example.com\",\"isPublic\":true,\"joinPolicy\"\
                    :\"INVITE\",\"followerCount\":42,\"iconUrl\":\"https://example.com/icon.png\"\
                    },\"resolved\":false,\"isSeriesPost\":false,\"pinnedAt\":null,\"\
                    isAuthor\":false,\"canDelete\":false,\"canComment\":true,\"createdAt\"\
                    :\"1970-01-01T00:00:00Z\",\"updatedAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"likeCount\":1,\"liked\":true,\"commentCount\":1,\"bookmarked\"\
                    :true,\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false}],\"preview\":\"This is a preview\
                    \ of the post content.\",\"previewMedia\":[{\"id\":1,\"url\":\"\
                    https://example.com/image.png\",\"fileName\":\"image.png\",\"\
                    fileSize\":102400,\"mediaType\":\"image/png\"}],\"mediaCount\"\
                    :1,\"wordCount\":120,\"coverImageUrl\":\"https://example.com/cover.png\"\
                    }]}"
...
  /projects/{handle}/posts:
    get:
      tags:
      - post
      summary: Get Posts By Project
      description: Get Posts By Project
      operationId: GetPostsByProject
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      - name: type
        in: query
        description: 게시글 타입
        required: false
        schema:
          type: string
      - name: authorHandle
        in: query
        description: 작성자 핸들
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPostsByProject:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
    post:
      tags:
      - post
      summary: Create Project Post
      description: 프로젝트에 글을 작성합니다.
      operationId: CreateProjectPost
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectPostRequest"
            examples:
              CreateProjectPost:
                value: "{\"title\":\"title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                  ,\"content\":{\"text\":\"This is a post content.\",\"json\":\"{\\\
                  \"text\\\": \\\"This is a post content.\\\"}\",\"mediaIds\":[1,2]},\"\
                  tags\":[\"java\",\"spring\"],\"projectTags\":[\"roadmap\"],\"referencedPostIds\"\
                  :[10,11],\"coverMediaId\":\"100\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                CreateProjectPost:
                  value: |-
                    {
                      "slug" : "slug"
                    }
...
  /projects/{handle}/posts/pinned:
    get:
      tags:
      - post
      summary: Get Pinned Posts By Project
      description: 프로젝트 대시보드에 고정된 게시글 목록을 조회합니다.
      operationId: GetPinnedPostsByProject
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostsResponse"
              examples:
                GetPinnedPostsByProject:
                  value: "{\"posts\":[{\"id\":1,\"slug\":\"test-slug\",\"title\":\"\
                    Test Post Title\",\"type\":\"SHORT\",\"status\":\"PUBLISHED\"\
                    ,\"scope\":\"WORKSPACE\",\"accessLevel\":\"FULL\",\"author\":{\"\
                    handle\":\"john-doe\",\"name\":\"John Doe\",\"profileImageUrl\"\
                    :\"https://example.com/john.png\"},\"project\":{\"handle\":\"\
                    my-project\",\"name\":\"나의 프로젝트\",\"description\":\"프로젝트 설명\"\
                    ,\"website\":\"https://example.com\",\"isPublic\":true,\"joinPolicy\"\
                    :\"INVITE\",\"followerCount\":42,\"iconUrl\":\"https://example.com/icon.png\"\
                    },\"resolved\":false,\"isSeriesPost\":false,\"pinnedAt\":null,\"\
                    isAuthor\":false,\"canDelete\":false,\"canComment\":true,\"createdAt\"\
                    :\"1970-01-01T00:00:00Z\",\"updatedAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"likeCount\":1,\"liked\":true,\"commentCount\":1,\"bookmarked\"\
                    :true,\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false}],\"preview\":\"This is a preview\
                    \ of the post content.\",\"previewMedia\":[{\"id\":1,\"url\":\"\
                    https://example.com/image.png\",\"fileName\":\"image.png\",\"\
                    fileSize\":102400,\"mediaType\":\"image/png\"}],\"mediaCount\"\
                    :1,\"wordCount\":120,\"coverImageUrl\":\"https://example.com/cover.png\"\
                    }]}"
...
  /tags/{tagId}/posts:
    get:
      tags:
      - post
      summary: Get Posts By Tag
      description: 태그가 붙은 공개 게시글 목록을 조회합니다.
      operationId: GetPostsByTag
      parameters:
      - name: tagId
        in: path
        description: 태그 ID
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      - name: type
        in: query
        description: 게시글 타입
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGetPostsResponse"
              examples:
                GetPostsByTag:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"title\":\"Test Post Title\"\
                    ,\"type\":\"SHORT\",\"status\":\"PUBLISHED\",\"scope\":\"WORKSPACE\"\
                    ,\"accessLevel\":\"FULL\",\"author\":{\"handle\":\"john-doe\"\
                    ,\"name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"project\":{\"handle\":\"my-project\",\"name\":\"나의 프로젝트\"\
                    ,\"description\":\"프로젝트 설명\",\"website\":\"https://example.com\"\
                    ,\"isPublic\":true,\"joinPolicy\":\"INVITE\",\"followerCount\"\
                    :42,\"iconUrl\":\"https://example.com/icon.png\"},\"resolved\"\
                    :false,\"isSeriesPost\":false,\"pinnedAt\":null,\"isAuthor\":false,\"\
                    canDelete\":false,\"canComment\":true,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"1970-01-01T00:00:00Z\",\"likeCount\":1,\"liked\"\
                    :true,\"commentCount\":1,\"bookmarked\":true,\"tags\":[{\"id\"\
                    :1,\"name\":\"java\",\"description\":\"자바 관련 태그\",\"postCount\"\
                    :10,\"followerCount\":3,\"projectHandle\":null,\"isFollowing\"\
                    :false}],\"preview\":\"This is a preview of the post content.\"\
                    ,\"previewMedia\":[{\"id\":1,\"url\":\"https://example.com/image.png\"\
                    ,\"fileName\":\"image.png\",\"fileSize\":102400,\"mediaType\"\
                    :\"image/png\"}],\"mediaCount\":1,\"wordCount\":120,\"coverImageUrl\"\
                    :\"https://example.com/cover.png\"}}]}}"
...
components:
  schemas:
...
    CreatePostResponse:
      title: CreatePostResponse
      required:
      - slug
      type: object
      properties:
        slug:
          type: string
          description: Post Slug
...
    CreatePostRequest:
      title: CreatePostRequest
      required:
      - content
      - type
      type: object
      properties:
        coverMediaId:
          type: string
          description: 커버 이미지로 사용할 미디어 ID (없으면 커버 없음)
          nullable: true
        referencedPostIds:
          type: array
          description: "이 글이 참조하는 게시글 ID 목록 (지정한 순서대로 노출, 최대 50개)"
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        title:
          type: string
          description: Title
          nullable: true
        content:
          required:
          - json
          - text
          type: object
          properties:
            json:
              type: string
              description: JSON Content
            mediaIds:
              type: array
              description: 사용된 미디어 ID 목록
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            text:
              type: string
              description: Text Content
          description: Content
        tags:
          type: array
          description: 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: string
          description: Post Status
          nullable: true
          enum:
          - DRAFT
          - PUBLISHED
...